support admin colorScheme in TextInput
